made kelly bigger started emu

+++ core/src/com/nrg/kelly/config/actors/Coordinates.java
@@ -1,22 +1,22 @@
 package com.nrg.kelly.config.actors;
 
 public abstract class Coordinates {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
